test: 불필요한 단위테스트 제거
